requirements.txt
asgiref==3.5.2
django-betterforms==2.0.0
django-betterforms3==2
crispy-bootstrap5==0.7
Django==4.1.2
django-crispy-forms==1.14.0
et-xmlfile==1.1.0
install==1.3.5
numpy==1.24.0
openpyxl==3.0.10
pandas==1.5.2
python-dateutil==2.8.2
pytz==2022.7
six==1.16.0
sqlparse==0.4.3
